apps/backend/requirements.txt
annotated-types==0.7.0
anyio==4.5.2
asgiref==3.8.1
backports.zoneinfo==0.2.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorlog==6.8.2
Deprecated==1.2.18
Django==4.2.19
djangorestframework==3.15.2
dnspython==2.6.1
docopt==0.6.2
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.6
fastapi-cli==0.0.6
ffmpeg-python==0.2.0
future==1.0.0
gevent==24.2.1
google-api-core==2.23.0
google-auth==2.36.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.0
idna==3.10
importlib_resources==6.4.4
iniconfig==2.0.0
Jinja2==3.1.4
limits==3.13.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==1.24.4
packaging==24.2
pandas==2.0.3
pluggy==1.5.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycountry==24.6.1
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
pysubs2==1.7.3
pytest==8.3.4
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.19
pythonanywhere==0.15.5
pythonanywhere-core==0.2.4
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
rich-toolkit==0.11.1
rsa==4.9
schema==0.7.7
shellingham==1.5.4
six==1.16.0
slowapi==0.1.9
snakesay==0.10.3
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.41.3
subs2cia==0.4.2
tabulate==0.9.0
tomli==2.2.1
tqdm==4.66.5
typer==0.15.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
wrapt==1.17.2
zipp==3.20.0
zope.event==5.0
zope.interface==7.0.1
